test(mariadb): limit clock stubs to synchronous calls
MariaDB integration tests kept a process-global clock stub active across awaited database work, so background tracer flushes could fail unrelated pool timing assertions.
